chore(core): save 20K by just importing inflate function
