+++ packages/botonic-core/src/hubtype-service.ts
@@ -1,9 +1,9 @@
 import axios, { AxiosResponse } from 'axios'
-import { inflate } from 'pako'
 import Pusher, { AuthOptions, Channel } from 'pusher-js'
 import Channels from 'pusher-js/types/src/core/channels/channels'
 
 import { Input, SessionUser } from './models'
+import { decompressData } from './pusher-utils'
 import { getWebpackEnvVar } from './utils'
 
 interface UnsentInput {
@@ -28,20 +28,6 @@ interface HubtypeServiceArgs {
   onEvent: any
   unsentInputs: () => UnsentInput[]
   server: ServerConfig
-}
-/**
- *
- * @param compressedData a string representing gzipped data previously encoded to base64 for very large contents
- * @returns a string representing the information of a very large content
- * Ref: https://stackoverflow.com/questions/4875020/javascript-decompress-inflate-unzip-ungzip-strings
- */
-function decompressData(compressedData) {
-  const strData = atob(compressedData)
-  const charData = strData.split('').map(x => x.charCodeAt(0))
-  const binData = new Uint8Array(charData)
-  const data = inflate(binData)
-  // @ts-ignore
-  return String.fromCharCode.apply(null, new Uint8Array(data))
 }
 
 const _WEBCHAT_PUSHER_KEY_ = getWebpackEnvVar(
